+ point update (not working)
